restructure file and add route for onbaording form
