+++ Frontend/src/App.jsx
@@ -6,23 +6,25 @@ import HomePage from "./Pages/LandingPage/HomePage";
 import ContactUs from "./Pages/ContactUs/ContactUs";
 import Login from "./Pages/Authentication/Login";
 import SignUp from "./Pages/Authentication/SignUp";
+import Onboarding from "./Pages/Onboarding/Onboarding";
 
 function App() {
-    return (
-        <BrowserRouter>
-            <Routes>
-                <Route path="/" element={<Layout />}>
-                    {/* landing page */}
-                    <Route index element={<HomePage />} />
-                    <Route path="/contact" element={<ContactUs />} />
-                    <Route path="*" element={<NotFound />} />
-                </Route>
-                {/* Authentication  */}
-                <Route path="/signup" element={<SignUp />} />
-                <Route path="/login" element={<Login />} />
-            </Routes>
-        </BrowserRouter>
-    );
+  return (
+    <BrowserRouter>
+      <Routes>
+        <Route path="/" element={<Layout />}>
+          {/* landing page */}
+          <Route index element={<HomePage />} />
+          <Route path="/contact" element={<ContactUs />} />
+          <Route path="*" element={<NotFound />} />
+        </Route>
+          {/* Authentication  */}
+          <Route path="/signup" element={<SignUp />} />
+          <Route path="/login" element={<Login />} />
+          <Route path="/onboarding" element={<Onboarding />} />
+      </Routes>
+    </BrowserRouter>
+  );
 }
 
 export default App;
