+++ Frontend/src/components/LandingPage/Home/Hero.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import heroImg from "../../../assets/images/hero-img.png";
-import ButtonSm from "../../utility/ButtonSm";
+import ButtonSm from "../../utility/Home/ButtonSm";
 import { useNavigate } from "react-router-dom";
 
 const Hero = () => {
